requirements.txt
fastapi
uvicorn[standard]
pydantic
numpy
pillow
transformers
python-dotenv
openai
replicate
redis
celery
sqlalchemy
alembic
pinecone
boto3
httpx
python-multipart
jinja2
